Added "logical item" column called item_number and necessary code.
Added functionality to dupe bundles and bitstreams for new versions of items. For now, we're duping it all.

The createNewVersion() method now will find the head revision of a set of revisions of an item and make the new revision replace the head, but have the properties of the parameter item.